RosterIQ/requirements.txt
fastapi
uvicorn
pandas
numpy
duckdb
plotly
openai
firebase-admin
python-dotenv
tavily-python
gunicorn
uvicorn
